refactor: move rotation-log readers from storage to node sync layer (#2266 step 2)
Per #2267 Option B, DAG-causal resolution belongs in the sync layer
(which owns the DAG), not in storage. Storage keeps the on-disk shape
and persistence (`load`, `save`, `append`); resolution moves out as a
pure-function module.

- New: `crates/node/src/sync/rotation_log_reader.rs` exposes
  `writers_at(&RotationLog, parents, happens_before)` and
  `latest_writers(&RotationLog)`. Both are pure (no I/O, no
  `StorageAdaptor`) — callers load via
  `calimero_storage::rotation_log::load::<S>(id)?` and pass `&log` in.
  Marked `#[allow(dead_code)]` until Step 3 wires them into
  `ContextStorageApplier::apply`.
- Removed from `crates/storage/src/rotation_log.rs`: `writers_at`,
  `latest_writers`, and the seven resolution-rule unit tests. The
  module's docstring is rewritten to make the storage/node split
  explicit.
- The four storage-side persistence tests (empty, append round-trip,
  idempotency, duplicate-delta hard error) stay put.
- The nine resolution-rule tests are migrated to the new node module
  with no behavior changes — just constructed via `RotationLog`
  literals instead of going through the mocked storage adaptor.

Storage rotation_log: 4 tests pass (persistence). Node
rotation_log_reader: 9 tests pass (ADR 0001 ordering rule). Workspace
builds clean.
